change binary data IO buffer size count from 4 byte to 8 byte

+++ MHD/C_src/CORE_C/read_psf_bin_data_c.h
@@ -21,7 +21,7 @@ struct psf_bin_work{
     
     int iflag_keep;
     int iflag_swap;
-    int lchar_out;
+    long lchar_out;
     long nprocs;
     long ilength;
